Count supported accelerometer motion interrupts in beacon formats
Configure the legacy 0.064 g motion threshold only when the selected Zephyr driver exposes a motion trigger, preserve its 32-bit count through DF5/7/8/C5 modulo encoding, and test trigger ordering and large-count rollover without inventing TAG wiring.

+++ nrf54_zephyr/src/formats.h
@@ -32,7 +32,7 @@ typedef struct {
     float luminosity_lux;
     float color_temp_k;
     uint16_t measurement_count;
-    uint16_t movement_count;
+    uint32_t movement_count; /* Legacy interrupt counter before format-specific modulo. */
     uint8_t motion_intensity;
     bool motion_detected;
     bool presence_detected;
